+++ crates/core/src/proto_config.rs
@@ -85,6 +85,17 @@ fn validate_reserved_words(
 }
 
 derive_enum!(
+    #[derive(ConfigEnum, Copy, Default)]
+    pub enum ConfigMode {
+        Global,
+        Local,
+        #[default]
+        Upwards,
+        UpwardsGlobal,
+    }
+);
+
+derive_enum!(
     #[derive(ConfigEnum, Default)]
     pub enum DetectStrategy {
         #[default]
@@ -570,6 +581,7 @@ impl ProtoConfigManager {
         })
     }
 
+    #[deprecated]
     pub fn get_global_config(&self) -> miette::Result<&ProtoConfig> {
         self.global_config.get_or_try_init(|| {
             debug!("Loading global config only");
@@ -578,6 +590,7 @@ impl ProtoConfigManager {
         })
     }
 
+    #[deprecated]
     pub fn get_local_config(&self, cwd: &Path) -> miette::Result<&ProtoConfig> {
         self.local_config.get_or_try_init(|| {
             debug!("Loading local config only");
@@ -591,6 +604,7 @@ impl ProtoConfigManager {
         })
     }
 
+    #[deprecated]
     pub fn get_merged_config(&self) -> miette::Result<&ProtoConfig> {
         self.all_config.get_or_try_init(|| {
             debug!("Merging loaded configs");
@@ -599,6 +613,7 @@ impl ProtoConfigManager {
         })
     }
 
+    #[deprecated]
     pub fn get_merged_config_without_global(&self) -> miette::Result<&ProtoConfig> {
         self.all_config_no_global.get_or_try_init(|| {
             debug!("Merging loaded configs without global");